More file reorganization.

+++ trunk/build/build.pl
@@ -1,12 +1,12 @@
 #!/usr/bin/perl
-
-# Portable version differences: Append "P" to BuildID. Different executable.
-
 #usage build.pl [build_settings.txt]
+
+$Debug_skip_compile = 1;
 
 use File::Spec;
 $TRUNK = File::Spec->rel2abs( __FILE__ );
 $TRUNK =~ s/[\\\/]build[\\\/][^\\\/]+$//;
+$TRUNK =~ s/\\/\//g;
 require "$TRUNK/build/script/common.pl";
 
 &Log("----> Reading control file.\n");
@@ -21,7 +21,7 @@ while(<SETF>) {
   elsif ($_ =~ /^#/) {next;}
   elsif ($_ =~ /^Set\s+(\S+)\s*=\s*(.*?)\s*$/i) {
     my $var=$1; my $val=$2;
-    if ($var =~ /^(.*?\.js\)\:/) {$Prefs{$var} = $val;}
+    if ($var =~ /^\(.*?\.js\)\:/) {$Prefs{$var} = $val;}
     else {$$1 = $2;}
   }
   else {&Log("ERROR: unhandled control file line $line: \"$_\"\n");}
@@ -31,82 +31,79 @@ $Xsprocess = $Executable;
 $Xsprocess =~ s/\.exe$/-srv.exe/i;
 @ModRepos = ($ModuleRepository1, $ModuleRepository2);
 
-&Log("----> Checking file paths.\n");
+&Log("----> Getting file paths.\n");
 if ($OutputDirectory =~ /^\./) {
   $OutputDirectory = File::Spec->rel2abs($OutputDirectory);
 }
-if ($ModuleRepository =~ /^\./) {
-  $ModuleRepository = File::Spec->rel2abs($ModuleRepository);
-}
-if (!-e $OutputDirectory) {
-  print "Create output directory \"$OutputDirectory\"?: [y/n]\n";
-  my $ans = <>;
-  if ($ans !~ /^y$/i) {exit;}
-  make_path($OutputDirectory);
-}
-if (!-e $ModuleRepository) {
-  &Log("WARNING: module repository does not exist: \"$ModuleRepository\"\n");
-}
-
-&Log("----> Deleting \"$TRUNK/$Name\"\n");
-if (-e "$TRUNK/$Name") {remove_tree("$TRUNK/$Name");}
-mkdir("$TRUNK/$Name");
-$DEVELOPMENT="$TRUNK/$Name/development";
-$INSTALLER="$TRUNK/$Name/installer";
-$FFEXTENSION="$TRUNK/$Name/extension";
-$PORTABLE="$TRUNK/$Name/portable";
+if (!-e $OutputDirectory) {make_path($OutputDirectory);}
+&Log("----> Deleting \"$TRUNK/build-files/$Name\"\n");
+if (-e "$TRUNK/build-files/$Name") {remove_tree("$TRUNK/build-files/$Name");}
+create_path("$TRUNK/build-files/$Name");
+$DEVELOPMENT="$TRUNK/build-files/$Name/development";
+$INSTALLER="$TRUNK/build-files/$Name/installer";
+$FFEXTENSION="$TRUNK/build-files/$Name/extension";
+$PORTABLE="$TRUNK/build-files/$Name/portable";
 if ("$^O" =~ /MSWin32/i) {$Appdata = `Set APPDATA`; $Appdata =~ s/APPDATA=(.*?)\s*$/$1/i;}
 else {&Log("ERROR: need to assign Linux application directory\n");}
 
 &writeCompileDeps();
 
 if ($MakeDevelopment =~ /true/i) {
-  &Log("----> BUILDING DEVELOPMENT ENVIRONMENT\n");
+  &Log("\n----> BUILDING DEVELOPMENT ENVIRONMENT\n");
   make_path($DEVELOPMENT);
-  &compileLibSword();
-  &writeAllPreferences($DEVELOPMENT);
+  &compileLibSword($DEVELOPMENT);
   my @manifest;
-  &copyExtensionFiles($DEVELOPMENT, @manifest, 1);
+  &copyExtensionFiles($DEVELOPMENT, \@manifest, 1);
+  &copyXulRunnerFiles($DEVELOPMENT);
+  &writeAllPreferences($DEVELOPMENT, \%Prefs, 1);
   &writeApplicationINI($DEVELOPMENT);
+  &compileStartup($DEVELOPMENT);
   &includeModules($IncludeModules, \@ModRepos, "$Appdata/$Vendor/$Name/Profiles/resources", $IncludeSearchIndexes);
   &writeManifest(\@manifest, $DEVELOPMENT);
+  &processLocales($IncludeLocales, "$DEVELOPMENT/defaults", "$Appdata/$Vendor/$Name/Profiles/*.default/extensions");
+  &writeRunScript($Name, $DEVELOPMENT, "dev");
 }
 if ($MakeFFextension =~ /true/i) {
-  &Log("----> BUILDING FIREFOX EXTENSION\n");
+  &Log("\n----> BUILDING FIREFOX EXTENSION\n");
   make_path($FFEXTENSION);
   &compileLibSword($FFEXTENSION);
-  &writeAllPreferences($FFEXTENSION);
   my @manifest;
-  &copyExtensionFiles($FFEXTENSION, @manifest);
+  &copyExtensionFiles($FFEXTENSION, \@manifest);
+  &writeAllPreferences($FFEXTENSION, \%Prefs);
   &writeFFInstallFiles();
   &writeManifest(\@manifest, $FFEXTENSION);
 }
 if ($MakePortable =~ /true/i) {
-  &Log("----> BUILDING PORTABLE VERSION\n");
+  &Log("\n----> BUILDING PORTABLE VERSION\n");
   make_path("$PORTABLE/$Name");
   make_path("$PORTABLE/resources");
   make_path("$PORTABLE/profile");
   &compileLibSword("$PORTABLE/$Name");
-  &writeAllPreferences("$PORTABLE/$Name");
   my @manifest;
-  &copyExtensionFiles("$PORTABLE/$Name", @manifest);
+  &copyExtensionFiles("$PORTABLE/$Name", \@manifest);
+  &copyXulRunnerFiles("$PORTABLE/$Name");
+  &writeAllPreferences("$PORTABLE/$Name", \%Prefs);
   &writeApplicationINI("$PORTABLE/$Name", 1);
   &compileStartup($PORTABLE);
   &includeModules($IncludeModules, \@ModRepos, "$PORTABLE/resoures", $IncludeSearchIndexes);
   &writeManifest(\@manifest, "$PORTABLE/$Name");
+  &processLocales($IncludeLocales, "$PORTABLE/$Name/defaults", "$PORTABLE/profile/extensions");
   open(NIN, ">:encoding(UTF-8)", "$PORTABLE/resources/newInstalls.txt") || die;
   print NIN "NewLocales;en-US\n"; # this opens language menu on first run
   close(NIN);
+  &createPortableZip($PORTABLE, "OutputDirectory/$Name-Portable-$Version/$Name Portable-$Version.zip");
+  &writeRunScript($Name, $PORTABLE, "port");
 }
 if ($MakeSetup =~ /true/i) {
   if ($Target eq "Windows") {
     if (-e "$XulswordExtras/installer") {
-      &Log("----> BUILDING SETUP INSTALLER\n");
+      &Log("\n----> BUILDING SETUP INSTALLER\n");
       make_path($INSTALLER);
       &compileLibSword($INSTALLER);
-      &writeAllPreferences($INSTALLER);
       my @manifest;
-      &copyExtensionFiles($INSTALLER, @manifest);
+      &copyExtensionFiles($INSTALLER, \@manifest);
+      &copyXulRunnerFiles($INSTALLER);
+      &writeAllPreferences($INSTALLER, \%Prefs);
       &writeApplicationINI($INSTALLER);
       &compileStartup($INSTALLER);
       if (-e "$Appdata/$Vendor/$Name/Profiles/resources/mods.d") {
@@ -117,18 +114,23 @@ if ($MakeSetup =~ /true/i) {
       }
       &includeModules($IncludeModules, \@ModRepos, "$Appdata/$Vendor/$Name/Profiles/resources", $IncludeSearchIndexes);
       &writeManifest(\@manifest, $INSTALLER);
+      &processLocales($IncludeLocales, "$INSTALLER/defaults", "$Appdata/$Vendor/$Name/Profiles/*.default/extensions");
       if (!-e "$XulswordExtras/installer/autogen") {mkdir "$XulswordExtras/installer/autogen";}
       &writeInstallerAppInfo("$XulswordExtras/installer/autogen/appinfo.iss");
       &writeInstallerLocaleinfo($IncludeLocales, "$XulswordExtras/installer/autogen/localeinfo.iss")
       &writeInstallerModuleUninstall($IncludeModules, $IncludeLocales, "$XulswordExtras/installer/autogen/uninstall.iss", "$Appdata/$Vendor/$Name/Profiles/resources");
+      &createSetupInstaller("$INSTALLER/scriptProduction.iss");
+      &writeRunScript($Name, $PORTABLE, "setup");
     }
     else {&Log("ERROR: the Inno Setup 5 installer script is missing.\n");}
   }
   else {&Log("ERROR: the installer has not yet been implemented for your platform.\n");}
 }
-&createLocaleExtensions(split(/\s*,\s*/, $IncludeLocales));
-if ("$^O" =~ /MSWin32/i && $Target eq "Windows") {&writeWindowsRegistryScript();}
-  
+
+&Log("FINISHED BUILDING\n");
+print "Press enter to close...\n";
+$p = <>;
+
 ################################################################################
 ################################################################################
 sub writeCompileDeps() {
@@ -147,6 +149,7 @@ sub writeCompileDeps() {
     print INFO "Set clucene=$CluceneSource\n";
     print INFO "Set sword=$SwordSource\n";
     print INFO "Set microsoftsdk=$MicrosoftSDK\n";
+    print INFO "Set Name=$Name\n";
     close(INFO);
   }
   else {&Log("ERROR: you need to pass path information to the compile script of your platform.\n");}
@@ -154,45 +157,60 @@ sub writeCompileDeps() {
 
 sub compileLibSword($) {
   my $do = shift;
-
   &Log("----> Compiling libsword binary.\n");
   if (!$CompiledAlready) {
     if ("$^O" =~ /MSWin32/i && $Target eq "Windows") {
-      if (!-e "$TRUNK/Cpp/cluceneMK/lib/Release/libclucene.lib") {
-        chdir("$TRUNK/Cpp/cluceneMK/lib");
-        `Compile.bat`;
+      if (!$Debug_skip_compile) {
+        if (!-e "$TRUNK/Cpp/cluceneMK/lib/Release/libclucene.lib") {
+          chdir("$TRUNK/Cpp/cluceneMK/lib");
+          `start "cluceneMK" /wait Compile.bat`;
+        }
+        if (!-e "$TRUNK/Cpp/swordMK/lib/Release/libsword.lib") {
+          chdir("$TRUNK/Cpp/swordMK/lib");
+          `start "swordMK" /wait Compile.bat`;
+        }
+        chdir("$TRUNK/Cpp");
+        if ($UseSecurityModule =~ /true/i) {
+          `start "xulsword" /wait Compile.bat`;
+        }
+        else {
+          `start "xulsword" /wait Compile.bat NOSECURITY`;
+        }
+        if (!-e "$TRUNK/Cpp/Release/xulsword.dll") {&Log("ERROR: libsword did not compile.\n");}
+        elsif (!-e "$do/xulsword.dll") {
+          copy("$TRUNK/Cpp/Release/xulsword.dll", $do);
+          $CompiledAlready = 1;
+        }
+        chdir("$TRUNK/build");
       }
-      if (!-e "$TRUNK/Cpp/swordMK/lib/Release/libsword.lib") {
-        chdir("$TRUNK/Cpp/swordMK/lib");
-        `Compile.bat`;
-      }
-      chdir("$TRUNK/Cpp");
-      if ($UseSecurityModule =~ /true/i) {
-        `Compile.bat`;
-      }
-      else {
-        `Compile.bat NOSECURITY`;
-      }
-      if (!-e "$TRUNK/Cpp/Release/xulsword.dll") {&Log("ERROR: libsword did not compile.\n");}
-      elsif ($do) {
-        copy("$TRUNK/Cpp/Release/xulsword.dll", $do);
-        $CompiledAlready = 1;
-      }
-      chdir("$TRUNK/build");
+      else {copy("$TRUNK/Cpp/Release/xulsword.dll", $do);}
     }
     else {&Log("ERROR: Please make and call a compile script for your platform.\n");}
   }
 }
 
-sub writeAllPreferences($) {
+sub writeAllPreferences($\%$) {
   my $do = shift;
-  &Log("----> Writing preferences.\n");
-  &writePrefs("$do/defaults/pref/prefs.js", \%Prefs);
-  &writePrefs("$do/defaults/pref/language.js", \%Prefs);
-  &writePrefs("$do/defaults/pref/buildprefs.js", \%Prefs);
+  my $pP = shift;
+  my $debug = shift;
 
-  if ($MakeDevelopment =~ /true/i) {
-    &writePrefs("$do/defaults/pref/debug.js", \%Prefs);
+  &Log("----> Writing preferences.\n");
+
+  foreach my $p (sort keys %{$pP}) {
+    if (!$pP->{$p}) {next;}
+    my $pref = $p;
+    if ($pref !~ s/^\((.*?)\)\://) {&Log("ERROR: malformed pref entry \"$pref\"\n"); next;}
+    my $fn = $1;
+    if ($fn eq "debug.js" && !$debug) {next;}
+    my $pfile = "$do/defaults/pref/$fn";
+    my $mode = ($prefFiles{$pfile} ? ">>":">").":encoding(UTF-8)";
+    open(PREF, $mode, $pfile) || die "Could not open pref file \"$pfile\"\n";
+    my $q = '"';
+    if ($pP->{$p} =~ s/^.*?(true|false).*?$/my $b=$1; $b=lc($b);/ie) {$q = "";}
+    if ($pref =~ /^HiddenTexts/) {$pP->{$p} =~ s/,/\;/; $pP->{$p} =~ s/\s+//; $pP->{$p}.=";"}
+    print PREF "pref(\"$pref\", $q".$pP->{$p}."$q);\n";
+    close(PREF);
+    $prefFiles{$pfile}++;
   }
 }
 
@@ -202,13 +220,12 @@ sub copyExtensionFiles($\@$) {
   my $makeDevelopment = shift;
 
   &Log("----> Copying Firefox extension files.\n");
-  &copy_dir("$TRUNK/xul/xulrunnerDevAndProd/xulrunner", $do, "\\.svn", "\\.svn");
+  &copy_dir("$TRUNK/xul/extension", $do, "\\.svn", "\\.svn");
 
-  &Log("----> Compiling manifest file.\n");
   if ($makeDevelopment) {
-    push(@{$manifestP}, "content xulsword file:../../xul/xulsword/");
-    push(@{$manifestP}, "locale xulsword en-US file:../../xul/en-US.xs/en-US-xulsword/xulsword/");
-    push(@{$manifestP}, "skin xulsword skin file:../../xul/skin/");
+    push(@{$manifestP}, "content xulsword file:../../../xul/xulsword/");
+    push(@{$manifestP}, "locale xulsword en-US file:../../../xul/en-US.xs/en-US-xulsword/xulsword/");
+    push(@{$manifestP}, "skin xulsword skin file:../../../xul/skin/");
   }
   else {
     push(@{$manifestP}, "content xulsword jar:chrome/xulsword.jar!/");
@@ -220,12 +237,20 @@ sub copyExtensionFiles($\@$) {
     &makeJAR("$do/chrome/en-US.xs.jar", "$TRUNK/xul/en-US.xs/en-US-xulsword");
     &makeJAR("$do/chrome/skin.jar", "$TRUNK/xul/skin");
   }
+  push (@{$manifestP}, "override chrome://global/locale/printPageSetup.dtd chrome://xulsword/locale/printPageSetup.dtd");
   push (@{$manifestP}, "override chrome://global/content/printPageSetup.xul chrome://xulsword/content/xulrunner/global/printPageSetup.xul");
   push (@{$manifestP}, "override chrome://global/content/printPreviewBindings.xml chrome://xulsword/content/xulrunner/global/printPreviewBindings.xml");
   push (@{$manifestP}, "override chrome://global/content/printPreviewProgress.xul chrome://xulsword/content/xulrunner/global/printPreviewProgress.xul");
   push (@{$manifestP}, "override chrome://global/content/printProgress.xul chrome://xulsword/content/xulrunner/global/printProgress.xul");
   push (@{$manifestP}, "override chrome://global/content/bindings/tree.xml chrome://xulsword/content/xulrunner/global/bindings/tree.xml");
   push (@{$manifestP}, "override chrome://mozapps/content/handling/dialog.xul chrome://xulsword/content/xulrunner/mozapps/handling/dialog.xul");
+}
+
+sub copyXulRunnerFiles($) {
+  my $do = shift;
+  &Log("----> Copying xulrunner files.\n");
+  copy_dir("$TRUNK/xulrunner", $do, "xulrunner\-stub\.exe", "");
+  move("$do/xulrunner.exe", "$do/$Xsprocess");
 }
 
 sub writeApplicationINI($$) {
@@ -248,15 +273,19 @@ sub writeApplicationINI($$) {
   close(INI);
 }
 
-sub &includeModules($\@$$) {
+sub includeModules($\@$$) {
   my @modules = split(/\s*,\s*/, shift);
   my $repsP = shift;
   my $do = shift;
   my $includeIndexes = shift;
+  
+  $includeIndexes = ($includeIndexes=~/true/i ? 1:0);
 
   my $path;
   foreach my $mod (@modules) {
     foreach my $di (@{$repsP}) {
+      #if ($di =~ /\$/) {$di =~ s/\\/\//g; print "HERE-> $di\n"; $di = eval($di);}
+      $di =~ s/(\$\w+)/my $e=$1; $e=eval($e)/eg;
       $path = "$di/mods.d/".lc($mod).".conf";
       if (-e $path) {last;}
     }
@@ -264,7 +293,7 @@ sub &includeModules($\@$$) {
     if (!$Conf{$mod}{"DataPath"}) {&getInfoFromConf($path);}
     if (!$Conf{$mod}{"DataPath"}) {&Log("ERROR: could not read conf file $path\n"); next;}
     my $mpath = $path;
-    $mpath =~ s/mods\.d\/[^\/]$//;
+    $mpath =~ s/mods\.d[\\\/][^\\\/]*$//;
     $mpath .= $Conf{$mod}{"DataPath"};
     if (!-e $mpath) {&Log("ERROR: module $mod files not found in $mpath.\n"); next;}
     if ($includeIndexes && !-e "$mpath/lucene") {&Log("ERROR: search index requested but is not available.\n");}
@@ -282,31 +311,37 @@ sub writeManifest(\@$) {
   my $maP = shift;
   my $od = shift;
   
-  open(MAN, ">:encoding(UTF-8)", "$od/chrome.manifest") || die "Could not open chrome manifest \"$od/chrome.manifest\"\n";
-  foreach $e (@{$maP}) {print MAN "$e\n";}
+  open(MAN, ">>:encoding(UTF-8)", "$od/chrome.manifest") || die "Could not open chrome manifest \"$od/chrome.manifest\"\n";
+  opendir(CHROME, "$od/chrome") || die "Could not open chrome directory \"$od/chrome\"\n";
+  @files = readdir(CHROME);
+  closedir(CHROME);
+  foreach my $file (@files) {
+    if ($file =~ /\.manifest/) {print MAN "manifest chrome/$file\n";}
+  }
+  foreach my $e (@{$maP}) {print MAN "$e\n";}
   close(MAN);
 }
 
 
 sub compileStartup($) {
   my $do = shift;
-
   my $rd;
   if ($do eq $PORTABLE)  {$rd = "runPortable";}
-  if ($do eq $INSTALLER) {$rd = "runMK";}
-  else {&Log("ERROR: no startup stub for \"$do\"\n"); return;}
+  else {$rd = "runMK";}
 
-  &Log("----> Compiling startup stub.\n");
-  chdir("$TRUNK/Cpp/$rd");
-  remove_tree("$TRUNK/Cpp/$rd/Release");
-  `Compile.bat`;
+  if (!$Debug_skip_compile) {
+    &Log("----> Compiling startup stub.\n");
+    remove_tree("$TRUNK/Cpp/$rd/Release");
+    chdir("$TRUNK/Cpp/$rd");
+    `start "$rd" /wait Compile.bat`;
+    chdir("$TRUNK/build");
+  }
   copy("$TRUNK/Cpp/$rd/Release/$rd.exe", "$do/$Executable");
-  move("$do/xulrunner.exe", "$do/$Xsprocess");
 }
 
 sub writeInstallerAppInfo($) {
   my $of = shift;
-  &Log("----> Writing installer script.\n");
+  &Log("----> Writing installer application info script.\n");
   open (ISS, ">:encoding(UTF-8)", $of) || die;
   print ISS "#define MyAppName \"$Name\"\n";
   print ISS "#define MyAppExeName \"$Executable\"\n";
@@ -325,9 +360,8 @@ sub writeInstallerAppInfo($) {
 sub writeInstallerLocaleinfo($$) {
   my @locales = split(/\s*,\s*/, shift);
   my $of = shift;
-
+  &Log("----> Writing installer locale script.\n");
   open(OUTF, ">:encoding(UTF-8)", $of) || die;
-  
   # this script use to write =false for all possible locales (needed??)
   foreach my $locale (@locales) {
     my $pl = $locale; $pl =~ s/-//g; # iss defines can't use "-"
@@ -341,8 +375,8 @@ sub writeInstallerModuleUninstall($$$$) {
   my @locales = split(/\s*,\s*/, shift);
   my $of = shift;
   my $md = shift;
-  
-  my $newInstalls = "NewLocales;" # don't list any locales- language menu should not open after setup install
+  &Log("----> Writing installer uninstall script.\n");
+  my $newInstalls = "NewLocales;"; # don't list any locales- language menu should not open after setup install
   $newInstalls .= "NewModules";
   if ($modules) {$newInstalls .= ";".join(";", @modules);}
 
@@ -356,12 +390,53 @@ sub writeInstallerModuleUninstall($$$$) {
   close(OUTF);
 }
 
+sub createSetupInstaller($) {
+  my $is = shift;
+  my $id = $is;
+  
+  &Log("----> Creating Windows setup installer.\n");
+  
+  $is =~ s/[\\\/][^\\\/]+$//;
+  if (!-e $is) {&Log("ERROR: installer script $is not found.\n"); return;}
+  chdir($id);
+  `"%ProgramFiles%\Inno Setup 5\ISCC.exe" "$is" > "$INSTALLER\autogen\setuplog.txt"`;
+  chdir("$TRUNK/build");
+}
+
+sub writeRunScript($$$) {
+  my $name = shift;
+  my $rd = shift;
+  my $type = shift;
+
+  &Log("----> Writing run script.\n");
+
+  my $s = "./run_$name-$type.pl";
+
+  if ($Target eq "Windows") {
+    open(SCR, ">:encoding(UTF-8)", $s) || die;
+    print SCR "#!/usr/bin/perl\n";
+    if ($type eq "dev") {
+      print SCR "chdir(\"".$DEVELOPMENT."\");\n";
+      print SCR "`$Name.exe -console -jsconsole`;\n";
+    }
+    elsif ($type eq "port") {
+      print SCR "`\"$PORTABLE\\$Name.exe\"`;\n";
+    }
+    elsif ($type eq "setup") {
+      &writeWindowsRegistryScript();
+      print SCR "`\"$INSTALLER\\$Name.exe\"`;\n";
+    }
+    close(SCR);
+  }
+  else {&Log("ERROR: no run scripts yet for your platform\n");}
+}
+
 sub writeWindowsRegistryScript() {
   &Log("----> Writing Windows registry script.\n");
   if (!-e "$TRUNK/build/autogen") {mkdir "$TRUNK/build/autogen";}
   open (REG, ">:encoding(UTF-8)", "$TRUNK/build/autogen/setRegistry.reg") || die;
   my $id = "$OutputDirectory/toCDROM/Install/setup"; $id =~ s/[\\\/]/\\\\/g;
-  my $rd = "$TRUNK/xulrunner"; $rd =~ s/[\\\/]/\\\\/g;
+  my $rd = "$INSTALLER"; $rd =~ s/[\\\/]/\\\\/g;
   my $ad = "$OutputDirectory/audio"; $ad =~ s/[\\\/]/\\\\/g;
   print REG "Windows Registry Editor Version 5.00\n";
   print REG "[HKEY_LOCAL_MACHINE\\SOFTWARE\\$Vendor\\$Name]\n";
@@ -371,14 +446,20 @@ sub writeWindowsRegistryScript() {
   print REG "\"Version\"=\"$Version\"\n";
 }
 
-sub createLocaleExtensions(\@) {
-  my @LocalesP = shift;
+sub processLocales($$$) {
+  my @locales = split(/\s*,\s*/, shift);
+  my $rdf = shift;
+  my $extdir = shift;
   
   &Log("----> Creating Locale extensions.\n");
-  foreach my $loc (@{$LocalesP}) {
+  foreach my $loc (@locales) {
     if ($loc eq "en-US") {next;}
-    &createLocaleExtension($loc, "$FFEXTENSION");
-    copy("$XulswordExtras/localeDev/$loc/$loc.rdf", "$FFEXTENSION/defaults");
+    &createLocaleExtension($loc, $extdir);
+    copy("$XulswordExtras/localeDev/$loc/$loc.rdf", $rdf);
   }
+}
+
+sub createLocaleExtension($$) {
+  &Log("WARNING: createLocaleExtension is not yet implemented\n");
 }
